requirements.txt
Flask==1.0.2
azure-keyvault>=0.3.6
